<refactor> : <MVC 모델의 의미와 맞게 재정립>
MVC 모델이 지향하는 바와 구현한 부분이 다르다고 생각되어 다시 기능들을 바꿨음
