+++ src/main/java/controller/RacingcarController.java
@@ -1,25 +1,45 @@
 package controller;
 
-import static view.Constant.INPUT_NUMBER;
-import static view.Constant.START_GAME;
-
-import camp.nextstep.edu.missionutils.Console;
 import model.RacingcarModel;
-import view.RacingcarView;
+import java.util.Arrays;
+import java.util.List;
+
 
 public class RacingcarController {
-    RacingcarView rv = new RacingcarView();
 
-    public void run(){
+    public void RacingStart(String car_names, String repeat){
         RacingcarModel rm;
-        String car_names;
-        String repeat_n;
+        List<String> car_names_list = SliceByComma(car_names);
+        ErrorCheck(car_names_list, repeat);
 
-        rv.Start();
-        car_names = Console.readLine();
-        rv.InputNuber();
-        repeat_n = Console.readLine();
-        rm = new RacingcarModel(car_names, repeat_n);
+        rm = new RacingcarModel(car_names_list, repeat);
     }
+    private void ErrorCheck(List<String> car_names_list, String repeat){
+
+        for(int i=0; i<car_names_list.size(); i++){
+            if(isCorrectLength(car_names_list.get(i))){
+                throw new IllegalArgumentException("자동차 이름 입력 길이가 잘못됨.");
+            }
+        }
+        if(!isNumeric(repeat)){
+            throw new IllegalArgumentException("반복횟수가 숫자가 아님.");
+        }
+    }
+    private boolean isCorrectLength(String car_name){
+        return (car_name.length() == 0 || car_name.length() > 5);
+    }
+
+    private boolean isNumeric(String user_number) {
+        return user_number.matches("\\d+");
+    }
+
+    private List<String> SliceByComma(String car_names){
+        car_names = car_names.replaceAll("\\s", "");
+        List<String> list = Arrays.asList(car_names.split(","));
+        return list;
+    }
+
+
+
 
 }
