+++ src/main/java/racingcar/Application.java
@@ -2,10 +2,11 @@ package racingcar;
 
 
 import controller.RacingcarController;
+import view.RacingcarView;
 
 public class Application {
     public static void main(String[] args) {
-        RacingcarController rc = new RacingcarController();
-        rc.run();
+        RacingcarView rv = new RacingcarView();
+        rv.run();
     }
 }
